requirements-dev.txt
-r requirements.txt
django-test-without-migrations==0.4
coverage